Moved ApiErrors to models.errors
