+++ app/uk/gov/hmrc/pillar2/connectors/AccountActivityConnector.scala
@@ -22,8 +22,8 @@ import uk.gov.hmrc.http.client.HttpClientV2
 import uk.gov.hmrc.http.given
 import uk.gov.hmrc.http.{HeaderCarrier, HttpResponse}
 import uk.gov.hmrc.pillar2.config.AppConfig
-import uk.gov.hmrc.pillar2.models.UnexpectedResponse
 import uk.gov.hmrc.pillar2.models.accountactivity.AccountActivityRequest
+import uk.gov.hmrc.pillar2.models.errors.UnexpectedResponse
 
 import scala.concurrent.{ExecutionContext, Future}
