+++ app/uk/gov/hmrc/pillar2/models/errors/ApiErrors.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.pillar2.models
+package uk.gov.hmrc.pillar2.models.errors
 
 import play.api.libs.json.{Json, OFormat}
