+++ app/uk/gov/hmrc/pillar2/models/errors/Pillar2ApiError.scala
@@ -18,7 +18,7 @@ package uk.gov.hmrc.pillar2.models.errors
 
 import play.api.libs.json.{Json, OFormat}
 
-case class Pillar2ApiError(code: String, message: String)
+final case class Pillar2ApiError(code: String, message: String)
 
 object Pillar2ApiError {
   given format: OFormat[Pillar2ApiError] = Json.format[Pillar2ApiError]
